Fix missing library in man page and minor doc typos
Add -lyaml to the link with line in the n2pkvna.3 man page.  Fix package
name typo in the RPM spec description section.  Correct a comment in
n2pkvna_hardware that claimed the minimum delay is 3ms when it's in
fact 1ms.  Fixed reversed bool and float in call to _n2pkvna_set_dds
where it disables the output.  Fixed transposed false and 0.0 in call
to _n2pkvna_set_dds where we disable the output.  Because of implicit
conversions, it was working correctly, but for the wrong reason.

No functional changes.