Clean up ranges and units handling in measurement screen
Modified the plot code in the measure screen to remember the plot
ranges and units for each base unit (frequency, impedance, admittance,
resistance, capacitance, etc.), and remember the ranges for dimensionless
quantities such as S parameters by parameter type and selected coodinate,
so that if you change to a different parameter type and back, the
previous settings are restored.  Some of the code for doing this was
already there, but incomplete.  Factored and simplified duplicated code.
Made handling of x, y and y2 axes more consistent.